requirements.txt
openai>=1.0.0
requests>=2.25.0
streamlit>=1.28.0
Pillow>=9.0.0 